Move workflow related code into separate module
